save history changes
